placeright-parser/requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.18
PyMuPDF==1.25.1
pdfplumber==0.11.4
spacy==3.8.3
scikit-learn==1.6.1
pydantic==2.10.4
python-dotenv==1.0.1
pytest==8.3.4
httpx==0.28.1
